feat: Refactor seeding to manual scripts
Replaces the automatic, frontend-triggered Netlify seeding function with two separate, manual Node.js scripts: one for seeding users and one for seeding application configuration.

The new user seeding script uses a complete, hardcoded list of users as requested.

Adds `seed:users` and `seed:config` commands to package.json for easy execution.

Removes the automatic seeding call from the frontend application.
