+++ package.json
@@ -8,7 +8,8 @@
     "build": "vite build",
     "lint": "eslint .",
     "preview": "vite preview",
-    "seed-users": "node scripts/seed-users.js"
+    "seed:users": "node scripts/seed-users.js",
+    "seed:config": "node scripts/seed-config.js"
   },
   "dependencies": {
     "@mdi/js": "^7.4.47",
